Decouple the bot functionality from the CLI
